# OpenAPI.yaml
openapi: 3.1.0
info:
  title: WebChatMIREA
  description: |-
    This is a draft API version of WebChat based on the OpenAPI 3.1
  version: 0.0.1
servers:
  - url: 'http/localhost:8080/'
...
paths:
...
  /ws/createChat:
    post:
      tags:
        - chat
      summary: Create chat
      description: ''
      operationId: CreateChat
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                  example: 1
                name:
                  type: string
                  example: test-chat
                group:
                  type: boolean
                  example: true
      responses:
        '201':
          description: Chat has been created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chat'
        '400':
          description: Failed to create room
...
components:
  schemas:
...
    Chat:
      type: object
      properties:
        id:
          type: string
          example: 1
        name:
          type: string
          example: test-chat
        clients:
          type: array
          items:
            $ref: '#/components/schemas/Client'
    Client:
      type: object
      properties:
        id:
          type: string
          example: 1
        chatID:
          type: string
          example: 1
        username:
          type: string
          example: john
...